+++ src/data/protocols/db/load-account-by-email-repository.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '../../domain/model/account'
+import { AccountModel } from '../../../domain/model/account'
 
 export interface LoadAccountByEmailRepository {
   load: (email: string) => Promise<AccountModel>
