+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../../domain/usecases/add-account'
 export * from '../../../domain/model/account'
-export * from '../../protocols/encrypter'
-export * from '../../../data/protocols/add-account-repository'
+export * from '../../protocols/criptography/encrypter'
+export * from '../../protocols/db/add-account-repository'
