+++ src/data/usecases/authentication/db-authentication.ts
@@ -1,4 +1,5 @@
-import { Authentication, AuthenticationModel } from '../../../domain/usecases/authentication'; import { LoadAccountByEmailRepository } from '../../protocols/load-account-by-email-repository'
+import { Authentication, AuthenticationModel } from '../../../domain/usecases/authentication'
+import { LoadAccountByEmailRepository } from '../../protocols/db/load-account-by-email-repository'
 
 export class DbAuthentication implements Authentication {
   private readonly loadAccountByEmailRepository: LoadAccountByEmailRepository
